requirements.txt
flask
pandas
finnhub-python
bokeh
python-dotenv
flask-cors
apscheduler
